backend/requirements.txt
# Core Framework
fastapi
uvicorn[standard]
pydantic
python-dotenv

# Database
psycopg2-binary

# Data Processing
pandas
numpy>=1.24.0,<2.0.0
pyarrow
fastparquet

# Mutual Fund Tools
mftool
yfinance

# PDF/Document Processing
pdfplumber
PyMuPDF
pypdf
python-multipart

# AI/ML
openai
faiss-cpu==1.7.4

# Google Cloud
google-cloud-storage==2.14.0

# Utilities
cachetools
deprecated
rapidfuzz
jinja2
itsdangerous
requests

# Auth
python-jose[cryptography]

# Testing
pytest